test(angular): update utils
